+++ app/controllers/staff/implementing_organisations_controller.rb
@@ -1,0 +1,29 @@
+class Staff::ImplementingOrganisationsController < Staff::BaseController
+  def new
+    @activity = Activity.find(params[:activity_id])
+    authorize @activity
+
+    @implementing_organisation = ImplementingOrganisation.new
+  end
+
+  def create
+    @activity = Activity.find(params[:activity_id])
+    authorize @activity
+
+    @implementing_organisation = ImplementingOrganisation.new(implementing_organisation_params)
+
+    if @implementing_organisation.valid?
+      @implementing_organisation.save
+      flash[:notice] = I18n.t("form.implementing_organisation.create.success")
+      redirect_to organisation_activity_path(@activity.organisation, @activity)
+    else
+      render :new
+    end
+  end
+
+  private
+
+  def implementing_organisation_params
+    params.require(:implementing_organisation).permit(:name, :organisation_type, :reference, :activity_id)
+  end
+end